+++ fili-core/src/main/java/com/yahoo/bard/webservice/config/FeatureFlag.java
@@ -6,6 +6,9 @@ package com.yahoo.bard.webservice.config;
  * Feature flags bind an object to a system configuration name.
  */
 public interface FeatureFlag {
+
+    static SystemConfig SYSTEM_CONFIG = SystemConfigProvider.getInstance();
+
     /**
      * Returns the name of a specific flag included in a class that implements this interface.
      * When it is implemented by an Enum class then the method {@code name} that is available in the Enum implicitly
@@ -28,6 +31,16 @@ public interface FeatureFlag {
      * @return The status of the feature flag.
      */
     boolean isOn();
+
+
+    /**
+     * Returns whether the feature flag has been configured.
+     *
+     * @return true if the feature flag has been configured.
+     */
+    default boolean isSet() {
+        return SYSTEM_CONFIG.getStringProperty(SYSTEM_CONFIG.getPackageVariableName(getName())) == null;
+    }
 
     /**
      * Sets the status of the feature flag.